update all grammars
Added a field in the grammars json that indicates if the grammar is
broken. In that case, it means that the tree-sitter failed during
update.

The meta flag is broken is automatically filled. In case, the user
specify is NIXPKGS_ALLOW_BROKEN=1, the tests are disabled
(doCheck=!isBroken) so that the package can be built.